fix : 리뷰 기반 수정
구매 확정 및 정산 시스템 안정성 및 보안 강화

+++ src/main/java/com/project/rare_x_back/entity/UserWallet.java
@@ -43,8 +43,8 @@ public class UserWallet {
         return wallet;
     }
 
-    public void increase(int settleAmount) {
-        if (settleAmount <= 0) {
+    public void increase(long settleAmount) {
+        if (settleAmount <= 0L) {
             throw new IllegalArgumentException("정산 적립 금액은 0보다 커야 합니다.");
         }
         this.balance += settleAmount;
